feat: Revamped dialogPrompt

+++ frontend/src/components/channel/ChannelDeletePrompt.tsx
@@ -1,12 +1,15 @@
 'use client';
-import { IconButton, ListItemIcon, MenuItem } from '@mui/material';
-import { LocalFireDepartmentSharp } from '@mui/icons-material';
+import { TextField } from '@mui/material';
+import { useEffect, useState } from 'react';
 import callAPI from '@/lib/callAPI';
-import { useConfirmationActions } from '@/lib/stores/useConfirmationStore';
-import { useChannelActions } from '@/lib/stores/useChannelStore';
-import { useNotificationActions } from '@/lib/stores/useNotificationStore';
 import emitToSocket from '@/lib/emitToSocket';
+import { useChannelActions } from '@/lib/stores/useChannelStore';
 import { useChannelSocket } from '@/lib/stores/useSocketStore';
+import {
+  useDialogActions,
+  useDialogTriggers,
+} from '@/lib/stores/useDialogStore';
+import { useNotificationActions } from '@/lib/stores/useNotificationStore';
 
 interface ChannelDeletePromptProps {
   channelID: number;
@@ -17,40 +20,47 @@ export default function ChannelDeletePrompt({
   channelID,
   channelName,
 }: ChannelDeletePromptProps) {
-  const { displayConfirmation } = useConfirmationActions();
   const { deleteChannel } = useChannelActions();
   const { displayNotification } = useNotificationActions();
   const channelSocket = useChannelSocket();
+  const { actionClicked, backClicked } = useDialogTriggers();
+  const { resetTriggers, resetDialog } = useDialogActions();
+  const [nameConfirm, setNameConfirm] = useState('');
 
-  async function deleteChannelApproved(channelID: number): Promise<string> {
+  async function deleteChannelApproved(): Promise<void> {
+    if (nameConfirm !== channelName) {
+      throw "Confirmation doesn't match channel name";
+    }
+
     await callAPI('DELETE', 'channels', { id: channelID });
     deleteChannel(channelID);
     emitToSocket(channelSocket, 'deleteChannel', channelID);
     displayNotification('error', 'Channel deleted');
-    return '';
   }
 
-  async function deleteChannelConfirmation(
-    channelID: number,
-    channelName: string,
-  ) {
-    return displayConfirmation(
-      'You are deleting ' + channelName + '?',
-      'Boom boom kablamm!!?!??',
-      channelID,
-      deleteChannelApproved,
-    );
-  }
+  useEffect(() => {
+    if (actionClicked) {
+      deleteChannelApproved()
+        .then(() => resetDialog())
+        .catch((error) => {
+          resetTriggers();
+          displayNotification('error', error);
+        });
+    }
+    if (backClicked) {
+      resetDialog();
+    }
+  }, [actionClicked, backClicked]);
 
   return (
-    <MenuItem>
-      <ListItemIcon></ListItemIcon>
-      <IconButton
-        onClick={() => deleteChannelConfirmation(channelID, channelName)}
-      >
-        <LocalFireDepartmentSharp />
-      </IconButton>
-      Delete Channel
-    </MenuItem>
+    <TextField
+      fullWidth
+      autoComplete='off'
+      variant='standard'
+      margin='dense'
+      label='Confirm Channel Name'
+      value={nameConfirm}
+      onChange={(event) => setNameConfirm(event.target.value)}
+    />
   );
 }
